Create a thermal_emission function combining line and continuum emission.
Combine line and continuum modules.
Add a thermal_emission function.  This now agrees with f_vth.pro from
SSW.
Extend continuum_emission function to handle multipleinput temperatures.
