+++ sunxspex/thermal.py
@@ -1,16 +1,111 @@
 import copy
 
-import numpy as np
 import astropy.units as u
-import scipy.interpolate
-import scipy.stats
+import numpy as np
+from scipy import interpolate, stats
+from sunpy.data import manager
 
-from sunxspex.io import load_chianti_lines_lite, load_xray_abundances
+from sunxspex.io import load_chianti_continuum, load_chianti_lines_lite, load_xray_abundances
 
-__all__ = ['line_emission']
+doc_string_params = """
+Parameters
+----------
+energy_edges: `astropy.units.Quantity`
+    The edges of the energy bins in a 1D N+1 quantity.
+
+temperature: `astropy.units.Quantity`
+    The temperature of the plasma.
+    Can be scalar or 1D of any length. If not scalar, the flux for each temperature
+    will be calculated. The first dimension of the output flux will correspond
+    to temperature.
+
+emission_measure: `astropy.units.Quantity`
+    The emission measure of the plasma at each temperature.
+    Must be same length as temperature or scalar.
+
+abundance_type: `str` (optional)
+    Abundance type to use.  Options are:
+        1. cosmic
+        2. sun_coronal - default abundance
+        3. sun_coronal_ext
+        4. sun_hybrid
+        5. sun_hybrid_ext
+        6. sun_photospheric
+        7. mewe_cosmic
+        8. mewe_solar
+    The values for each abundance type is stored in the global
+    variable DEFAULT_ABUNDANCES which is generated by `setup_default_abundances`
+    function. To load different default values for each abundance type,
+    see the docstring of that function.
+
+relative_abundances: `tuple` of `tuples` of (`int`, `float`) (optional)
+    The relative abundances of different elements as a fraction of their
+    default abundances defined by abundance_type.
+    Each tuple represents the values for a given element.
+    The first entry represents the atomic number of the element.  If this is not
+    entered as an int, it is rounded to the nearest int.
+    The second entry represents the axis represents the fraction by which the 
+    element's default abundance should be scaled.
+
+observer_distance: `astropy.units.Quantity` (Optional)
+    The distance between the source and the observer. Scales output to observer distance
+    and unit by 1/length.
+    Default=1 AU.
+
+Returns
+-------
+flux: `astropy.units.Quantity`
+    The photon flux as a function of temperature and energy.
+"""
+
+def setup_continuum_parameters(filename=None):
+    """
+    Define continuum intensities as a function of temperature.
+
+    Intensities are set as global variables and used in
+    calculation of spectra by other functions in this module. They are in
+    units of per volume emission measure at source, i.e. they must be
+    divided by 4 * pi R**2 to be converted to physical values where
+    R**2 is observer distance.
+
+    Intensities are derived from output from the CHIANTI atomic physics database.
+    The default CHIANTI data used here is collected from
+    `https://hesperia.gsfc.nasa.gov/ssw/packages/xray/dbase/chianti/chianti_cont_1_250_v71.sav`.
+    This includes contributions from thermal bremsstrahlung and tw-photon interactions.
+    To use a different file, provide the URL/file location via the filename kwarg,
+    e.g. to include only thermal bremsstrahlung, set the filename kwarg to 
+    'https://hesperia.gsfc.nasa.gov/ssw/packages/xray/dbase/chianti/chianti_cont_1_250_v70_no2photon.sav'
+
+    Parameters
+    ----------
+    filename: `str` (optional)
+        URL or file location of the CHIANTI IDL save file to be used.
+    """
+    global _CONT_T_K,_CONT_LOG10T, _CONT_T_KEV, _CONT_ELEMENT_IDX, _CONT_SORTED_ELEMENT_INDEX, wavelength, _CONT_E_KEV, _CONT_SPECTRAL_BIN_WIDTHS_KEV, _CONT_INTENSITY_PER_EM_AT_SOURCE, _CONT_INTENSITY_UNIT
+    if filename:
+        with manager.override_file("chianti_continuum", uri=filename):
+            cont_info = load_chianti_continuum()
+    else:
+        cont_info = load_chianti_continuum()
+    _CONT_ELEMENT_IDX = cont_info.element_index.data
+    _CONT_SORTED_ELEMENT_INDEX = np.sort(_CONT_ELEMENT_IDX)
+
+    T_grid = cont_info.temperature.data * cont_info.attrs["units"]["temperature"]
+    _CONT_T_K = T_grid.to_value(u.K)
+    _CONT_LOG10T = np.log10(_CONT_T_K)
+    _CONT_T_KEV = T_grid.to_value(u.keV, equivalencies=u.temperature_energy())
+
+    wavelength = cont_info.wavelength.data * cont_info.attrs["units"]["wavelength"]
+    dwave_AA = (cont_info.attrs["wavelength_edges"][1:] -
+                cont_info.attrs["wavelength_edges"][:-1]).to_value(u.AA)
+    _CONT_E_KEV = wavelength.to_value(u.keV, equivalencies=u.spectral())
+    _CONT_SPECTRAL_BIN_WIDTHS_KEV = _CONT_E_KEV * dwave_AA / wavelength.to_value(u.AA)
+
+    _CONT_INTENSITY_PER_EM_AT_SOURCE = cont_info.data * 4 * np.pi  # Convert from per sterradian to emission at source.
+    _CONT_INTENSITY_UNIT = cont_info.attrs["units"]["data"] * u.sr  # Remove per sterradian in accordance with above scaling.
 
 
-def define_line_parameters(filename=None):
+def setup_line_parameters(filename=None):
     """Define line intensities as a function of temperature for calculating line emission.
 
     Line intensities are set as global variables and used in the
@@ -45,7 +140,7 @@ def define_line_parameters(filename=None):
     _LINE_ATOMIC_NUMBERS = line_info.atomic_number.data
 
 
-def define_default_abundances(filename=None):
+def setup_default_abundances(filename=None):
     """
     Read default abundance values into global variable.
 
@@ -65,10 +160,181 @@ def define_default_abundances(filename=None):
     else:
         DEFAULT_ABUNDANCES = load_xray_abundances()
 
-
 # Read line, continuum and abundance data into global variables.
-define_line_parameters()
-define_default_abundances()
+setup_continuum_parameters()
+setup_line_parameters()
+setup_default_abundances()
+
+
+@u.quantity_input(energy_edges=u.keV,
+                  temperature=u.K,
+                  emission_measure=(u.cm**(-3), u.cm**(-5)),
+                  observer_distance=u.cm)
+def thermal_emission(energy_edges,
+                     temperature,
+                     emission_measure,
+                     abundance_type="sun_coronal",
+                     relative_abundances=None,
+                     observer_distance=(1*u.AU).to(u.cm)):
+    f"""Calculate the thermal X-ray emission (lines + continuum) from the solar atmosphere.
+
+    The emission is calculated as a function of temperature and emission measure.
+    Which continuum mechanisms are included --- free-free, free-bound, or two-photon --- are
+    determined by the file from which the comtinuum parameters are loaded.
+    To change the file used, see the setup_continuum_parameters() function.
+    
+    {doc_string_params}"""
+    energy_edges_keV = energy_edges.to_value(u.keV)
+    temperature_K = temperature.to_value(u.K)
+    if temperature.isscalar:
+        temperature_K = np.array([temperature_K])
+    continuum_flux = _continuum_emission(energy_edges_keV, temperature_K, abundance_type,
+                                         relative_abundances)
+    line_flux = _line_emission(energy_edges_keV, temperature_K, abundance_type, relative_abundances)
+    flux = ((continuum_flux + line_flux) * emission_measure / 
+                (4 * np.pi * observer_distance**2))
+    if temperature.isscalar and emission_measure.isscalar:
+        flux = flux[0]
+    return flux
+
+
+@u.quantity_input(energy_edges=u.keV,
+                  temperature=u.K,
+                  emission_measure=(u.cm**(-3), u.cm**(-5)),
+                  observer_distance=u.cm)
+def continuum_emission(energy_edges,
+              temperature,
+              emission_measure,
+              abundance_type="sun_coronal",
+              relative_abundances=None,
+              observer_distance=(1*u.AU).to(u.cm)):
+    f"""Calculate the thermal X-ray continuum emission from the solar atmosphere.
+
+    The emission is calculated as a function of temperature and emission measure.
+    Which continuum mechanisms are included --- free-free, free-bound, or two-photon --- are
+    determined by the file from which the comtinuum parameters are loaded.
+    To change the file used, see the setup_continuum_parameters() function.
+
+    {doc_string_params}"""
+    temperature_K = temperature.to_value(u.K)
+    if temperature.isscalar:
+        temperature_K = np.array([temperature_K])
+    flux = _continuum_emission(energy_edges.to_value(u.keV), temperature_K,
+                               abundance_type=abundance_type,
+                               relative_abundances=relative_abundances)
+    flux *= emission_measure / (4 * np.pi * observer_distance**2)
+    if temperature.isscalar:
+        flux = flux[0]
+    return flux
+
+def _continuum_emission(energy_edges_keV,
+                        temperature_K,
+                        abundance_type,
+                        relative_abundances):
+    """
+    Calculates emission-measure-normalized X-ray continuum emission at the source.
+
+    Output must be multiplied by emission measure and divided by 4*pi*observer_distance**2
+    to get physical values.
+    Which continuum mechanisms are included --- free-free, free-bound, or two-photon --- are
+    determined by the file from which the comtinuum parameters are loaded.
+    To change the file used, see the setup_continuum_parameters() function.
+    
+    Parameters
+    ----------
+    energy_edges_keV: 1-D array-like
+        Boundaries of contiguous spectral bins in units on keV.
+
+    temperature_K: 1-D array-like
+        The temperature(s) of the plasma in unit of K.  Must not be a scalar.
+
+    abundance_type: `str`
+        Set of default abundances to be used.  See the continuum_emission function docstring
+        for options.
+
+    relative_abundances: same as for continuum_emission function.
+    """
+    # Handle inputs and derive some useful parameters from them
+    log10T_in = np.log10(temperature_K)
+    T_in_keV = temperature_K / 11604518  # Convert temperature from K to keV.
+    # Get energy bins centers based on geometric mean.
+    energy_gmean_keV = stats.gmean(np.vstack((energy_edges_keV[:-1], energy_edges_keV[1:])))
+
+    #####  Calculate Abundances #####
+    # Calculate abundance of each desired element.
+    default_abundances = DEFAULT_ABUNDANCES[abundance_type].data
+    n_abundances = len(default_abundances)
+    rel_abund_values = np.ones(n_abundances)
+    if relative_abundances:
+        # Convert input relative abundances to array where
+        # first axis is atomic number, i.e == index + 1
+        # Second axis is relative abundance value.
+        rel_abund_array = np.array(relative_abundances).T
+        rel_idx = np.rint(rel_abund_array[0]).astype(int) - 1
+        rel_abund_values[rel_idx] = rel_abund_array[1]
+    abundance_mask = np.zeros(n_abundances, dtype=bool)
+    abundance_mask[_CONT_ELEMENT_IDX] = True
+    abundances = default_abundances * rel_abund_values * abundance_mask
+
+    #####  Calculate Continuum Intensity Summed Over All Elements
+    #####  For Each Temperature as a function of Energy/Wavelength ######
+    # Before looping over temperatures, let's perform the calculations that are
+    # used over again in the for loop.
+    # 1. If many temperatures are input, convolve intensity grid with abundances for all
+    # temperatures here.  If only a few temperatures are input, do this step only
+    # when looping over input temperatures.  This minimizes computation.
+    n_tband = 3
+    n_t_grid = len(_CONT_LOG10T)
+    n_temperature_K = len(temperature_K)
+    n_thresh = n_temperature_K * n_tband
+    if n_thresh >= n_t_grid:
+        print("All Ts multiplied by abundance.")
+        intensity_per_em_at_source_allT = np.zeros(_CONT_INTENSITY_PER_EM_AT_SOURCE.shape[1:])
+        for i in range(0, n_t_grid):
+            intensity_per_em_at_source_allT[i] = np.matmul(
+                abundances[_CONT_SORTED_ELEMENT_INDEX],
+                _CONT_INTENSITY_PER_EM_AT_SOURCE[:, i])
+    # 2. Add dummy axes to energy and temperature grid arrays for later vectorized operations.
+    repeat_E_grid = _CONT_E_KEV[np.newaxis, :]
+    repeat_T_grid = _CONT_T_KEV[:, np.newaxis]
+    dE_grid_keV = _CONT_SPECTRAL_BIN_WIDTHS_KEV[np.newaxis, :]
+    # 3. Indentify the indices of the temperature bins containing each input temperature and
+    # the bins above and below them.  For each input temperature, these three bins will
+    # act as a temperature band over which we'll interpolate the continuum emission.
+    selt = np.digitize(log10T_in, _CONT_LOG10T) - 1
+    tband_idx = selt[:, np.newaxis] + np.arange(n_tband)[np.newaxis, :]
+
+    # Finally, loop over input temperatures and calculate continuum emission for each.
+    flux = np.zeros((n_temperature_K, len(energy_gmean_keV)))
+    for j, logt in enumerate(log10T_in):
+        # If not already done above, calculate continuum intensity summed over
+        # all elements as a function of energy/wavelength over the temperature band.
+        if n_thresh < n_t_grid:
+            element_intensities_per_em_at_source = _CONT_INTENSITY_PER_EM_AT_SOURCE[:, tband_idx[j]]
+            intensity_per_em_at_source = np.zeros(element_intensities_per_em_at_source.shape[1:])
+            for i in range(0, n_tband):
+                intensity_per_em_at_source[i] = np.matmul(
+                    abundances[_CONT_SORTED_ELEMENT_INDEX],
+                    element_intensities_per_em_at_source[:, i])
+        else:
+            intensity_per_em_at_source = intensity_per_em_at_source_allT[tband_idx[j]]
+
+        ##### Calculate Continuum Intensity at Input Temperature  ######
+        ##### Do this by interpolating the normalized temperature component
+        ##### of the intensity grid to input temperature(s) and then rescaling.
+        # Calculate normalized temperature component of the intensity grid.
+        exponent = (repeat_E_grid / repeat_T_grid[tband_idx[j]])
+        exponential = np.exp(np.clip(exponent, None, 80))
+        gaunt = intensity_per_em_at_source / dE_grid_keV * exponential
+        # Interpolate the normalized temperature component of the intensity grid the the
+        # input temperature.
+        flux[j] = _interpolate_continuum_intensities(
+            gaunt, _CONT_LOG10T[tband_idx[j]], _CONT_E_KEV, energy_gmean_keV, logt)
+    # Rescale the interpolated intensity.
+    flux = flux * np.exp(-(energy_gmean_keV[np.newaxis, :] / T_in_keV[:, np.newaxis]))
+
+    # Put intensity into correct units.
+    return flux * _CONT_INTENSITY_UNIT
 
 
 @u.quantity_input(energy_edges=u.keV,
@@ -76,71 +342,32 @@ define_default_abundances()
                   emission_measure=(u.cm**(-3), u.cm**(-5)),
                   observer_distance=u.cm)
 def line_emission(energy_edges,
-                  temperature,
-                  emission_measure,
-                  abundance_type="sun_coronal",
-                  relative_abundances=None,
-                  observer_distance=(1*u.AU).to(u.cm)):
+         temperature,
+         emission_measure,
+         abundance_type="sun_coronal",
+         relative_abundances=None,
+         observer_distance=(1*u.AU).to(u.cm)):
     """
     Calculate thermal line emission from the solar corona.
 
-    The emission is calculated as a function of temperature and emission measure.
-
-    Parameters
-    ----------
-    energy_edges: `astropy.units.Quantity`
-        The edges of the energy bins in a 1D N+1 quantity.
-
-    temperature: `astropy.units.Quantity`
-        The temperature of the plasma.
-        Can be scalar or 1D of any length. If not scalar, the flux for each temperature
-        will be calculated. The first dimension of the output flux will correspond
-        to temperature.
-
-    emission_measure: `astropy.units.Quantity`
-        The emission measure of the plasma at each temperature.
-        Must be same length as temperature or scalar.
-
-    abundance_type: `str` (optional)
-        Abundance type to use.  Options are:
-            1. cosmic
-            2. sun_coronal - default abundance
-            3. sun_coronal_ext
-            4. sun_hybrid
-            5. sun_hybrid_ext
-            6. sun_photospheric
-            7. mewe_cosmic
-            8. mewe_solar
-        The values for each abundance type is stored in the global
-        variable DEFAULT_ABUNDANCES which is generated by `define_default_abundances`
-        function. To load different default values for each abundance type,
-        see the docstring of that function.
-
-    relative_abundances: `tuple` of `tuples` of (`int`, `float`) (optional)
-        The relative abundances of different elements as a fraction of their
-        default abundances defined by abundance_type.
-        Each tuple represents the values for a given element.
-        The first entry represents the atomic number of the element.  If this is not
-        entered as an int, it is rounded to the nearest int.
-        The second entry represents the axis represents the fraction by which the 
-        element's default abundance should be scaled.
-
-    observer_distance: `astropy.units.Quantity` (Optional)
-        The distance between the source and the observer. Scales output to observer distance
-        and unit by 1/length.
-        Default=1 AU.
-
-    Returns
-    -------
-    flux: `astropy.units.Quantity`
-        The photon flux as a function of temperature and energy.
-    """
+    {docstring_params}"""
     # For ease of calculation, convert inputs to known units and structures.
     energy_edges_keV = energy_edges.to_value(u.keV)
-    n_energy_bins = len(energy_edges_keV)-1
     temperature_K = temperature.to_value(u.K)
     if temperature.isscalar:
         temperature_K = np.array([temperature_K])
+    flux = _line_emission(energy_edges_keV, temperature_K, abundance_type, relative_abundances)
+    flux *= emission_measure / (4 * np.pi * observer_distance**2)    
+    if temperature.isscalar and emission_measure.isscalar:
+        flux = flux[0]
+    return flux
+
+
+def _line_emission(energy_edges_keV,
+                   temperature_K,
+                   abundance_type,
+                   relative_abundances):
+    n_energy_bins = len(energy_edges_keV)-1
     n_temperatures = len(temperature_K)
 
     # Find indices of lines within user input energy range.
@@ -193,19 +420,48 @@ def line_emission(energy_edges,
         # Use binned_statistic to determine which spectral bins contain
         # components of line emission and sum over those line components
         # to get the total emission is each spectral bin.
-        flux = scipy.stats.binned_statistic(line_spectrum_bins, split_line_intensities,
-                                            "sum", n_energy_bins, (0, n_energy_bins-1)).statistic
+        flux = stats.binned_statistic(line_spectrum_bins, split_line_intensities,
+                                      "sum", n_energy_bins, (0, n_energy_bins-1)).statistic
     else:
         flux = np.zeros((n_temperatures, n_energy_bins))
     
     # Scale flux by observer distance, emission measure and spectral bin width
     # and put into correct units.
     energy_bin_widths = (energy_edges_keV[1:] - energy_edges_keV[:-1]) * u.keV
-    flux = (flux * _LINE_INTENSITY_UNIT * emission_measure /
-            (energy_bin_widths * 4 * np.pi * observer_distance**2))
-    if temperature.isscalar and emission_measure.isscalar:
-        flux = flux[0]
+    flux = (flux * _LINE_INTENSITY_UNIT / energy_bin_widths)
 
+    return flux
+
+
+def _interpolate_continuum_intensities(data_grid, log10T_grid, energy_grid_keV, energy_keV, log10T):
+    # Determine valid range based on limits of intensity grid's spectral extent
+    # and the normalized temperature component of intensity.
+    n_tband = len(log10T_grid)
+    vrange, = np.where(data_grid[0] > 0)
+    for i in range(1, n_tband):
+        vrange_i, = np.where(data_grid[i] > 0)
+        if len(vrange) < len(vrange_i):
+            vrange = vrange_i
+    data_grid = data_grid[:, vrange]
+    energy_grid_keV = energy_grid_keV[vrange]
+    energy_idx, = np.where(energy_keV < energy_grid_keV.max())
+
+    # Interpolate temperature component of intensity and derive continuum intensity.
+    flux = np.zeros(energy_keV.shape)
+    if len(energy_idx) > 0:
+        energy_keV = energy_keV[energy_idx]
+        cont0 = interpolate.interp1d(energy_grid_keV, data_grid[0])(energy_keV)
+        cont1 = interpolate.interp1d(energy_grid_keV, data_grid[1])(energy_keV)
+        cont2 = interpolate.interp1d(energy_grid_keV, data_grid[2])(energy_keV)
+        # Calculate the continuum intensity as the weighted geometric mean
+        # of the interpolated values across the temperature band of the
+        # temperature component of intensity.
+        logelog10T = np.log(log10T)
+        x0, x1, x2 = np.log(log10T_grid)
+        flux[energy_idx]  = np.exp(
+            np.log(cont0) * (logelog10T - x1) * (logelog10T - x2) / ((x0 - x1) * (x0 - x2)) +
+            np.log(cont1) * (logelog10T - x0) * (logelog10T - x2) / ((x1 - x0) * (x1 - x2)) +
+            np.log(cont2) * (logelog10T - x0) * (logelog10T - x1) / ((x2 - x0) * (x2 - x1)) )
     return flux
 
 
@@ -266,7 +522,7 @@ def _calculate_abundance_normalized_line_intensities(logT, data_grid, line_logT_
         # Interpolate the 2nd axis to produce a function that gives the data
         # as a function of 1st axis, say energy, at a given value along the 2nd axis,
         # say "temperature".
-        get_intensities_at_logT = scipy.interpolate.interp1d(line_logT_bins[indx], data_grid[:, indx], kind="quadratic")
+        get_intensities_at_logT = interpolate.interp1d(line_logT_bins[indx], data_grid[:, indx], kind="quadratic")
         # Use function to get interpolated_data as a function of the first axis at
         # the input value along the 2nd axis,
         # e.g. line intensities as a function of energy at a given temperature.
