--- sunxspex/thermal/tests/__init__.py
@@ -1,4 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""
-This module contains package tests.
-"""
